LPS-127106 Move resolveEditableValue to utils/editable-value
