+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/components/fragment-content/FragmentContent.test.js
@@ -25,10 +25,10 @@ import {
 } from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/Controls';
 import {EditableProcessorContextProvider} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/EditableProcessorContext';
 import FragmentContent from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/FragmentContent';
-import resolveEditableValue from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/resolveEditableValue';
 import {BACKGROUND_IMAGE_FRAGMENT_ENTRY_PROCESSOR} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/backgroundImageFragmentEntryProcessor';
 import {VIEWPORT_SIZES} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/config/constants/viewportSizes';
 import {StoreAPIContextProvider} from '../../../../../src/main/resources/META-INF/resources/page_editor/app/store';
+import resolveEditableValue from '../../../../../src/main/resources/META-INF/resources/page_editor/app/utils/editable-value/resolveEditableValue';
 
 jest.mock(
 	'../../../../../src/main/resources/META-INF/resources/page_editor/app/services/serviceFetch',
@@ -36,7 +36,7 @@ jest.mock(
 );
 
 jest.mock(
-	'../../../../../src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/resolveEditableValue',
+	'../../../../../src/main/resources/META-INF/resources/page_editor/app/utils/editable-value/resolveEditableValue',
 	() => jest.fn(() => Promise.resolve(['Default content']))
 );
