+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/components/fragment-content/resolveEditableValue.test.js
@@ -13,7 +13,7 @@
  * details.
  */
 
-import resolveEditableValue from '../../../../../src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/resolveEditableValue';
+import resolveEditableValue from '../../../../../src/main/resources/META-INF/resources/page_editor/app/utils/editable-value/resolveEditableValue';
 
 jest.mock(
 	'../../../../../src/main/resources/META-INF/resources/page_editor/app/config',
